+++ src/main.cpp
@@ -53,6 +53,7 @@
 // ------------------MACROS, NAMESPACES AND DEFINITIONS--------------------
 // ------------------------------------------------------------------------
 
+#define CLI
 #ifndef JSON              // Checking if JSON is defined or not
 nlohmann::json cloudSave; // Cloud save is loaded in the memory once the user signs in
 nlohmann::json localSave; // Local save is also loaded in the memory once the user signs in
@@ -1240,104 +1241,9 @@ bool generateKey()
     return true;
 }
 
-// ------------------------------------------------------------------------
-// --------------------------CLI FUNCTIONALITY-----------------------------
-// ------------------------------------------------------------------------
-
-namespace cli
-{
-    void test(std::vector<std::string> args)
-    {
-        tabulate::Table t;
-        t.add_row({"Starting test mode (only for dev-builds)"});
-        t.add_row({"Only call it when you know what you are doing"});
-        t.add_row({"Arguments got: "});
-        for (std::string arg : args)
-            t.add_row({arg});
-        t[0][0].format().font_color(tabulate::Color::blue).font_style({tabulate::FontStyle::bold}).font_align(tabulate::FontAlign::center);
-        t[1][0].format().font_color(tabulate::Color::red).font_style({tabulate::FontStyle::bold}).font_align(tabulate::FontAlign::center);
-        t[2][0].format().font_color(tabulate::Color::green);
-        std::cout << t << std::endl;
-    }
-    void version(std::vector<std::string> args)
-    {
-        if (args.size() > 1)
-        {
-            std::transform(args[1].begin(), args[1].end(), args[1].begin(), [](char &a)
-                           { return std::tolower(a); });
-            if (args[1] == "classic" || args[1] == "simple")
-            {
-                std::cout << "MarkIt!" << std::endl;
-                std::cout << "Version: " << APP_VERSION << std::endl;
-            }
-            else if (args[1] == "modern")
-            {
-                tabulate::Table t;
-                t.add_row({"MarkIt!", "", ""});
-                t.add_row({"", "Version", APP_VERSION});
-                for (int i = 0; i < 3; i++)
-                {
-                    t[0][i].format().font_align(tabulate::FontAlign::center).font_color(tabulate::Color::yellow).font_style({tabulate::FontStyle::bold});
-                    t[1][i].format().font_align(tabulate::FontAlign::center).font_color(tabulate::Color::yellow).font_style({tabulate::FontStyle::bold});
-                }
-                t[0][1].format().font_background_color(tabulate::Color::yellow);
-                t[0][2].format().font_background_color(tabulate::Color::yellow);
-                t[1][0].format().font_background_color(tabulate::Color::yellow);
-                std::cout << t << std::endl;
-            }
-            else
-            {
-                std::cout << "No format \"" << args[1] << "\" known" << std::endl;
-            }
-        }
-        else
-        {
-            tabulate::Table t;
-            t.add_row({"MarkIt!", "", ""});
-            t.add_row({"", "Version", APP_VERSION});
-            for (int i = 0; i < 3; i++)
-            {
-                t[0][i].format().font_align(tabulate::FontAlign::center).font_color(tabulate::Color::yellow).font_style({tabulate::FontStyle::bold});
-                t[1][i].format().font_align(tabulate::FontAlign::center).font_color(tabulate::Color::yellow).font_style({tabulate::FontStyle::bold});
-            }
-            t[0][1].format().font_background_color(tabulate::Color::yellow);
-            t[0][2].format().font_background_color(tabulate::Color::yellow);
-            t[1][0].format().font_background_color(tabulate::Color::yellow);
-            std::cout << t << std::endl;
-        }
-    }
-    void display(std::vector<std::string> args)
-    {
-        if (!exist(storageFile))
-        {
-            std::cout << "No data found" << std::endl;
-            return;
-        }
-        tabulate::Table t;
-        json j = json::parse(_read_from_file(storageFile));
-        std::vector<todo> temp = j["data"];
-        t.add_row({"Name", "Description", "Time", "Completed"});
-        for (int i = 0; i < temp.size(); i++)
-        {
-            if (temp[i].isComplete)
-            {
-                t.add_row({temp[i].name, temp[i].desc, temp[i].time, "Yes"});
-                t[i + 1][3].format().font_color(tabulate::Color::green);
-            }
-            else
-            {
-                t.add_row({temp[i].name, temp[i].desc, temp[i].time, "No"});
-                t[i + 1][3].format().font_color(tabulate::Color::red);
-            }
-        }
-        for (int i = 0; i < 4; i++)
-        {
-            t.column(i).format().font_align(tabulate::FontAlign::center);
-            t[0][i].format().font_color(tabulate::Color::blue).font_style({tabulate::FontStyle::bold});
-        }
-        std::cout << t << std::endl;
-    }
-}
+#ifdef CLI
+#include "cli.hpp"                      // For adding the CLI functionality
+#endif
 
 // ------------------------------------------------------------------------
 // ----------------------------MAIN FUNCTION-------------------------------
@@ -1348,6 +1254,7 @@ int main(int argc, char *argv[])
     std::ios_base::sync_with_stdio(false);
     std::cin.tie(nullptr);
     std::cout.tie(nullptr);
+#ifdef CLI
     if (argc > 1)
     {
         std::vector<std::string> args(0);
@@ -1367,6 +1274,7 @@ int main(int argc, char *argv[])
         }
         return 0;
     }
+#endif
     set_title();
     initscr();
     cbreak();
