+++ tools/clean.cpp
@@ -10,8 +10,8 @@ int main()
 {
 #ifdef _WIN32
     system("cls");
-    system("del *.o >nul 2>&1");
-    system("del *.exe >nul 2>&1");
+    system("del src\\*.o >nul 2>&1");
+    system("del src\\*.exe >nul 2>&1");
     system("del bin\\*.dat >nul 2>&1");
     system("del bin\\main.exe >nul 2>&1");
     system("del bin\\main_debug.exe >nul 2>&1");
@@ -25,6 +25,7 @@ int main()
 #else
     system("clear");
     system("rm *.exe > /dev/null 2>&1");
+    system("rm src/*.exe > /dev/null 2>&1");
     system("rm bin/*.dat > /dev/null 2>&1");
     system("rm bin/main.exe > /dev/null 2>&1");
     system("rm bin/main_debug.exe > /dev/null 2>&1");
@@ -32,13 +33,14 @@ int main()
     system("rm *.o > /dev/null 2>&1");
     system("rm *.dat > /dev/null 2>&1");
     system("rm main_debug > /dev/null 2>&1");
-    system("rm main > /dev/null 2>&1");
+    system("rm src/main > /dev/null 2>&1");
+    system("rm tools/main > /dev/null 2>&1");
     system("rm build > /dev/null 2>&1");
     system("rm generates > /dev/null 2>&1");
     system("rm clean > /dev/null 2>&1");
 #endif
 #ifdef _WIN32
-    if (!exist("bin/main.exe") && !exist("bin/main_debug.exe") && !exist("main") && !exist("main_debug") && !exist("main.o") && !exist("main.out") && !exist("main_debug.out") && !exist("clean") && !exist("build") && !exist("build.exe") && !exist("data.dat") && !exist("bin/data.dat"))
+    if (!exist("bin/main.exe") && !exist("bin/main_debug.exe") && !exist("src/main") && !exist("main_debug") && !exist("main.o") && !exist("main.out") && !exist("main_debug.out") && !exist("clean") && !exist("build") && !exist("build.exe") && !exist("data.dat") && !exist("bin/data.dat"))
     {
         cout << "\u001b[32m   ________                          __\u001b[0m\n";
         cout << "\u001b[32m  / ____/ /__  ____ _____  ___  ____/ /\u001b[0m\n";
@@ -73,7 +75,7 @@ int main()
         cout << "/_/ |_/\\____/\\__/   \\____/_/\\___/\\__,_/_/ /_/\\___/\\__,_/\n\n";
     }
 #else
-    if (!exist("bin/main.exe") && !exist("bin/main_debug.exe") && !exist("main") && !exist("main_debug") && !exist("main.o") && !exist("main.out") && !exist("main_debug.out") && !exist("generates") && !exist("clean") && !exist("clean.exe") && !exist("build") && !exist("build.exe") && !exist("data.dat") && !exist("bin/data.dat"))
+    if (!exist("bin/main.exe") && !exist("bin/main_debug.exe") && !exist("src/main") && !exist("main_debug") && !exist("main.o") && !exist("main.out") && !exist("main_debug.out") && !exist("generates") && !exist("clean") && !exist("clean.exe") && !exist("build") && !exist("build.exe") && !exist("data.dat") && !exist("bin/data.dat"))
     {
         cout << "\u001b[32m   ________                          __\u001b[0m\n";
         cout << "\u001b[32m  / ____/ /__  ____ _____  ___  ____/ /\u001b[0m\n";
