+++ tools/prereq.bat
